refactor(cli): extract command-safety-manifest derivation into command_contract (#4255)
